update jsdoc generate folder
