+++ .gitignore
@@ -7,3 +7,4 @@
 /InternalContractAbi
 docs/apiv2
 .vscode
+apidoc
