+++ scripts/jsdoc-to-md.js
@@ -1,7 +1,7 @@
 /* eslint-disable import/no-extraneous-dependencies */
 const jsdoc2md = require('jsdoc-to-markdown');
 const fs = require('fs');
-// const { replaceImport } = require('@conflux-dev/jsdoc-tsimport-plugin');
+const { replaceTSImport } = require('./replaceTSimport');
 
 const DOC_FOLDER = './docs/api/';
 
@@ -13,33 +13,9 @@ const files = [{
   name: 'PoS.md',
 }];
 
-/**
- * A regex to capture all doc comments.
- */
-const docCommentsRegex = /\/\*\*\s*(?:[^\*]|(?:\*(?!\/)))*\*\//g;
-
-/**
- * Finds a ts import.
- */
-const importRegex = /import\(['"](\@?[\.\/_a-zA-Z0-9-\$]*)(?:\.js)?['"]\)\.?([_a-zA-Z0-9-\$]*)?/g;
-
-function replaceTSImport(e) {
-  return e.source.replace(docCommentsRegex,
-    substring => {
-      return substring.replace(importRegex,
-        (_substring2, relImportPath, symbolName) => {
-        // const moduleId = getModuleId(e.filename, relImportPath);
-          const moduleId = null;
-          return (moduleId) ? `module:${moduleId}${symbolName ? `~${symbolName}` : ''}` : symbolName;
-        });
-    });
-}
-
 async function renderFile(fileMeta) {
   let source = fs.readFileSync(fileMeta.source, 'utf8');
-  source = replaceTSImport({
-    source,
-  });
+  source = replaceTSImport({ source });
   const result = await jsdoc2md.render({
     // files: fileMeta.source,
     source,
@@ -54,5 +30,3 @@ async function main() {
 }
 
 main().catch(console.error);
-
-
